requirements.txt
streamlit
pandas
sqlalchemy
pymysql
bcrypt
plotly
openai
streamlit-aggrid
Pillow
python-dotenv
cloud-sql-python-connector[pymysql]